refactor: 가게 웨이팅 취소 비즈니스 정책의 책임 도메인으로 이동
