+++ fooding-core/src/main/java/im/fooding/core/model/waiting/StoreWaiting.java
@@ -1,5 +1,7 @@
 package im.fooding.core.model.waiting;
 
+import im.fooding.core.global.exception.ApiException;
+import im.fooding.core.global.exception.ErrorCode;
 import im.fooding.core.model.BaseEntity;
 import im.fooding.core.model.store.Store;
 import jakarta.persistence.Column;
@@ -92,6 +94,10 @@ public class StoreWaiting extends BaseEntity {
     }
 
     public void cancel() {
+        if (status != StoreWaitingStatus.WAITING) {
+            throw new ApiException(ErrorCode.STORE_WAITING_ILLEGAL_STATE_CANCEL);
+        }
+
         this.status = StoreWaitingStatus.CANCELLED;
     }
