+++ fooding-core/src/test/java/im/fooding/core/service/waiting/StoreWaitingServiceTest.java
@@ -228,23 +228,4 @@ class StoreWaitingServiceTest extends TestConfig {
         Assertions.assertThat(storeWaiting.getStatus())
                 .isEqualTo(StoreWaitingStatus.CANCELLED);
     }
-
-    @Test
-    @DisplayName("웨이팅 상태가 아닌 경우 웨이팅을 취소할 수 없다.")
-    public void testCancel_fail_whenStatusIsNotWaiting() {
-        // given
-        Store store = storeRepository.save(StoreDummy.create());
-        WaitingUser user = waitingUserRepository.save(WaitingUserDummy.create(store));
-
-        StoreWaiting storeWaiting = storeWaitingRepository.save(StoreWaitingDummy.create(user, store));
-        storeWaiting.cancel();
-
-        // when & then
-        ApiException e = assertThrows(
-                ApiException.class,
-                () -> storeWaitingService.cancel(storeWaiting.getId())
-        );
-        Assertions.assertThat(e.getErrorCode())
-                .isEqualTo(ErrorCode.STORE_WAITING_ILLEGAL_STATE_CANCEL);
-    }
 }
